services/requirements.txt
# Harbor - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# AI
anthropic>=0.39.0

# HTTP Requests (for integrations)
requests>=2.31.0

# File Watching
watchdog>=4.0.0

# Task Scheduling
schedule>=1.2.0

# Environment Variables
python-dotenv>=1.0.0

# Document Text Extraction
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
